+++ smartyard/db/__init__.py
@@ -1,0 +1,6 @@
+from .database import create_db_connection
+from .temps import Temps
+from .users import Users
+
+
+__all__ = ["create_db_connection", "Temps", "Users"]
